Removed a superfluous cast. Trying 'l' for Win64.

+++ src/logic/map_objects/immovable.cc
@@ -1228,7 +1228,7 @@ PlayerImmovable::PlayerImmovable(const MapObjectDescr& mo_descr)
 PlayerImmovable::~PlayerImmovable() {
 	if (workers_.size())
 		log("PlayerImmovable::~PlayerImmovable: %" PRIuS " workers left!\n",
-		    static_cast<long unsigned int>(workers_.size()));
+		    workers_.size());
 }
 
 /**
